work towards getting oss-fuzz to work again
